tests/requirements.txt
pytest==7.4.3
httpx==0.25.2
fakeredis==2.20.0
psycopg2-binary==2.9.9
pandas==2.1.2
